d/t/docker-in-lxd: Match "file:/" instead of "file:///" on copy_local_apt_files.
The command "apt-get indextargets" returns a URI field with "file:/"
as the prefix.  We've been trying to match "file:///", which is
sometimes unsuccessful and leads to errors because the local apt cache
is not copied over to the container.  We can use "file:/" instead,
which will work on both scenarios.  It's also not a problem to just
strip the "file:" prefix, because even paths like "///var/cache" are
correctly resolved by the filesystem.